+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## 0.6.3 - [2020-XX-XX]
+### Added
+- Added tests for `allowed_principals_with` and `non_whitelisted_allowed_principals`
 ### Fixes
 - Fix types in `allowed_principals_with`, `non_whitelisted_allowed_principals` and `PSEUDO_PARAMETERS`.